+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add &lt;temperature_feels_like&gt; tag
 - BREAKING: Replace --current-city CLI argument with &lt;city&gt; format tag
+- BREAKING: Exit after printing by specifying negative interval instead of no interval
 
 ## 1.1.1 - 2020-06-15